refactor: 홈 화면 정렬 Select 컴포넌트 네이밍 변경
