+++ components/pages/Home/SortDirectionSelect/SortDirectionSelect.stories.tsx
@@ -1,0 +1,9 @@
+import type { Meta, StoryFn } from "@storybook/react";
+import { SortDirectionSelect } from "./SortDirectionSelect";
+
+export default {
+  title: "UI/SortDirectionSelect",
+  component: SortDirectionSelect,
+} as Meta;
+
+export const Default: StoryFn<typeof SortDirectionSelect> = (args) => <SortDirectionSelect {...args} />;
